Pass explicitly `bootstrapper` to `Broker`
The `CliqueInfo` message that was sent to the parent was never received
in the test and thus causes an unhandled message warning.

Adding a `bootstrapper` actor ref gives us more freedom. First to
properly test that the message is sent, which is one of the goal of
this actor and also to be less tight to the parent, if latter we want
the `Broker` actor to directly send this message to the `CliqueManager`
for example.

Related to: #53