Move `Storages` creation from `PlatformConfig` to `Mode`
IMHO DB creation should be handle at the root of the project, so we can
handle it (create/close/etc) and not in the config, where I expect to
have only static values defined in `*.conf` files.

WARNING: `test` doesn't compile yet, latter commit will clean the way we
create our DB as well as `Storages`, it will be easier then to handle
the tests.