Fix `listWallets` that was returning a `WalletLocked`
The bug was that we were calling the `checkIsMiner` function when
listing the wallets while it's not needed and this function requires the wallets to be
unlocked.

In the fix we reduce the usage of the `checkIsMiner` function only to
the ones that really needed to know if they have a miner wallet or not.

Fix #128