Show double unlock edge case in the test
This edge case happen when the wallet have to be lazily-automatically-locked
and we call the unlock on that wallet.

In such case we are locking the secretStorage and not updating the `lastAccess` time.
The wallet is then unlocked in the service and return true.
But on the next call, for example to get balance,
as the access time was not updated, we re-lock the wallet.